+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.2.0-beta.0](https://github.com/Seungwoo321/genai-yt/compare/v0.1.1...v0.2.0-beta.0) (2026-03-08)
+
+### Features
+
+* **transcript:** add speech-to-text fallback with yt-dlp and whisper-cpp ([53cbd64](https://github.com/Seungwoo321/genai-yt/commit/53cbd641f4f3abf994d2ef8fc534a7aaf6d15e04))
+
+### Documentation
+
+* add speech-to-text fallback documentation ([1f567db](https://github.com/Seungwoo321/genai-yt/commit/1f567db4e5866646729cc4bf0ee4cc87cecb92c8))
+
 ## [0.1.1](https://github.com/Seungwoo321/genai-yt/compare/v0.1.1-beta.1...v0.1.1) (2026-03-08)
 
 ## [0.1.1-beta.1](https://github.com/Seungwoo321/genai-yt/compare/v0.1.1-beta.0...v0.1.1-beta.1) (2026-03-08)